Move applicationUtils to the utils/application folder
This mirrors what we do for assessments
