+++ server/controllers/apply/applicationsController.test.ts
@@ -14,10 +14,10 @@ import Apply from '../../form-pages/apply'
 
 import paths from '../../paths/apply'
 import { DateFormats } from '../../utils/dateUtils'
-import { firstPageOfApplicationJourney, getResponses, isUnapplicable } from '../../utils/applicationUtils'
+import { firstPageOfApplicationJourney, getResponses, isUnapplicable } from '../../utils/applications/utils'
 
 jest.mock('../../utils/validation')
-jest.mock('../../utils/applicationUtils')
+jest.mock('../../utils/applications/utils')
 jest.mock('../../services/tasklistService')
 
 describe('applicationsController', () => {
