+++ server/services/assessmentService.test.ts
@@ -11,12 +11,12 @@ import { getBody, updateAssessmentData } from '../form-pages/utils'
 import TasklistPage, { TasklistPageInterface } from '../form-pages/tasklistPage'
 import { DataServices, TaskListErrors } from '../@types/ui'
 import { ValidationError } from '../utils/errors'
-import { ApplicationOrAssessmentResponse, getResponseForPage, getResponses } from '../utils/applicationUtils'
+import { ApplicationOrAssessmentResponse, getResponseForPage, getResponses } from '../utils/applications/utils'
 
 jest.mock('../data/assessmentClient.ts')
 jest.mock('../data/personClient.ts')
 jest.mock('../form-pages/utils')
-jest.mock('../utils/applicationUtils')
+jest.mock('../utils/applications/utils')
 
 describe('AssessmentService', () => {
   const assessmentClient = new AssessmentClient(null) as jest.Mocked<AssessmentClient>
