+++ server/utils/assessments/utils.ts
@@ -14,7 +14,7 @@ import { format, differenceInDays, add, getUnixTime } from 'date-fns'
 import { ApprovedPremisesAssessment as Assessment, ApprovedPremisesApplication } from '@approved-premises/api'
 import { tierBadge } from '../personUtils'
 import { DateFormats } from '../dateUtils'
-import { getArrivalDate, getResponseForPage } from '../applicationUtils'
+import { getArrivalDate, getResponseForPage } from '../applications/utils'
 import paths from '../../paths/assess'
 import { TasklistPageInterface } from '../../form-pages/tasklistPage'
 import Assess from '../../form-pages/assess'
